Updated routes

+++ server/db/models/product.js
@@ -25,7 +25,8 @@ const Product = db.define('product', {
     type: Sequelize.STRING
   },
   price: {
-    type: Sequelize.DECIMAL(10, 2)
+    //IN PENNIES!!!!!!!
+    type: Sequelize.INTEGER
   },
   size: {
     type: Sequelize.INTEGER
